Convert Java sources to C++
Rewrite the Java sources to C++ by simply rewriting the syntax to the
similar C++ syntax and adding dummy classes/functions for the required
J2ME classes as well as the Java StringBuilder class.

Add a basic CMakeLists.txt for building the code.

This compiles and links but nothing works.