Use SFML for drawing graphics
Add a wrapper around the J2ME graphics functionality using SFML for
drawing.

This only implements what is needed for this game but hopefully wraps
that mostly correctly although some fairly nasty hacks are required
because of the different APIs.

Removed use of threads since that is completely unnecessary and can
only introduce very hard to fix issues.