Implement font drawing
Add support for using fonts instead of bitmap for drawing text. This
looks much, much nicer.

Use the Skranji Font for the actual font which seems to fit pretty
well into the games theme and is distributed under Open Font License
which should be compatible with the GPL used by this game.