Don't treat compiler warnings as errors by default
Seems like SFML is using the same global CMake option to set warnings
as errors which causes the build to fail on Windows because of some
warnings when building SFML on that platform.

We should change it so we don't use the same option name but for now
just set it to false by default.